manifests: Present initial manifest
The auto plugin was checked manually by:
Manually creating a db.dns via the fedora container.
Exposing the secondary-dns pod.
Running nslookup once the auto plugin refreshed the database.

The RBACs were not tested, but they are a skeleton that can
be easily updated once we have the watcher logic.

Signed-off-by: Or Shoval <oshoval@redhat.com>