controller: wait for kubevirt before starting
If the controller will be starting before the virtualmachineinstances
crd will be available the controller enter to a retry loop with a backoff
and eventually crash.
Once virtualmachineinstances crd will be available, the backoff time may
be very high which will cause a big delay in the controller startup.

To prevent this, the manager will wait for the crd to appear before
starting the controller.

Signed-off-by: Alona Paz <alkaplan@redhat.com>