cnao: Use cached version
Use the cached CNAO version, without CR.
The CR is deployed afterwards according the local CR that which we have
in the KubeSecondaryDNS repository itself (deploys multus only).

It saves time because the images are already cached,
and also eliminates the need to update CNAO versions,
as bumping kubevirtci will take care about it.

In case the provider is external, CNAO CR won't be deployed.
The cluster owner should have multus installed.

Signed-off-by: Or Shoval <oshoval@redhat.com>