coreDNS: Bump version
Since we moved from k8s.gcr.io to registry.k8s.io
the current version is aliased in those two.
This cause CNAO update test to fail because it expects
the new URL but the image is detected as the previous one.

Having a new image will solve this, and it should be fine to update
the version every now and then.

Signed-off-by: Or Shoval <oshoval@redhat.com>